updated delete identities task

+++ src/database/types.rs
@@ -16,7 +16,8 @@ pub struct RecoveryEntry {
     pub new_commitment:      Hash,
 }
 
+#[derive(Hash, PartialEq, Eq)]
 pub struct DeletionEntry {
-    pub leaf_index: i64, // TODO: consider using a larger type for leaf indices
+    pub leaf_index: usize,
     pub commitment: Hash,
 }
